requirements-build.txt
click
pandas
pyarrow
python-decouple
datetime
seaborn
matplotlib
